feat(v0.3.0): implement product returns, float inventory migration, supplier CRUD, global confirm UI, and general QoL enhancements

+++ backend/app/schemas/batch.py
@@ -12,5 +12,5 @@ class BatchOut(BaseModel):
     batch_code: str
     expiry_date: date
     import_price: Decimal
-    quantity_remaining: int
+    quantity_remaining: float
     created_at: datetime
